docs: Replace Jekyll-based site with MkDocs v0.15.3
As an attempt to improve documentation for PerfKit Benchmarker, this
restructures the docs website using a the MkDocs framework. The MkDocs
framework was created for Docs websites and makes it easier for other
to extend and contribute to the documentation. It also includes Search
support.

Additionally, this commit breaks up the documentation into sub-directories,
as well simplifies the 'Getting Started' docs.